Switched to vite based No-Bundler solution
