+++ README.md
@@ -1,14 +1,16 @@
-[devStatus]: https://img.shields.io/badge/Website-Under%20Development-blue?style=flat-square&logo=visualstudiocode&color=007ACC&logoColor=007ACC
+[devstatus]: https://img.shields.io/badge/Website-Under%20Development-blue?style=flat-square&logo=visualstudiocode&color=007ACC&logoColor=007ACC
 [website]: https://suveer.netlify.app/
-[Netlify-Status]: https://api.netlify.com/api/v1/badges/d1b32dce-27bf-4c3f-b185-d582887a0982/deploy-status
-[Netlify-Deploys]: https://app.netlify.com/sites/suveer/deploys
+[netlify-status]: https://api.netlify.com/api/v1/badges/d1b32dce-27bf-4c3f-b185-d582887a0982/deploy-status
+[netlify-deploys]: https://app.netlify.com/sites/suveer/deploys
 
 # Suveer R 👋
+
 [![devStatus]][website]
 
-[![Netlify-Status]][Netlify-Deploys]
+[![Netlify-Status]][netlify-deploys]
 
 ### Game Developer | Frontend programmer
+
 ```
 C#
 Java
