+++ client.d.ts
@@ -1,0 +1,51 @@
+/* eslint-disable no-unused-vars */
+declare module "~pages" {
+  import type { RouteRecordRaw } from "vue-router"
+  const routes: RouteRecordRaw[]
+  export default routes
+}
+
+declare module "pages-generated" {
+  import type { RouteRecordRaw } from "vue-router"
+  const routes: RouteRecordRaw[]
+  export default routes
+}
+
+declare module "virtual:generated-pages" {
+  import type { RouteRecordRaw } from "vue-router"
+  const routes: RouteRecordRaw[]
+  export default routes
+}
+
+declare module "virtual:pwa-register" {
+  export type RegisterSWOptions = {
+    immediate?: boolean
+    onNeedRefresh?: () => void
+    onOfflineReady?: () => void
+    onRegistered?: (registration: ServiceWorkerRegistration | undefined) => void
+    onRegisterError?: (error: any) => void
+  }
+
+  export function registerSW(
+    options?: RegisterSWOptions
+  ): (reloadPage?: boolean) => Promise<void>
+}
+
+declare module "virtual:pwa-register/vue" {
+  // @ts-ignore ignore when vue is not installed
+  import type { Ref } from "vue"
+
+  export type RegisterSWOptions = {
+    immediate?: boolean
+    onNeedRefresh?: () => void
+    onOfflineReady?: () => void
+    onRegistered?: (registration: ServiceWorkerRegistration | undefined) => void
+    onRegisterError?: (error: any) => void
+  }
+
+  export function useRegisterSW(options?: RegisterSWOptions): {
+    needRefresh: Ref<boolean>
+    offlineReady: Ref<boolean>
+    updateServiceWorker: (reloadPage?: boolean) => Promise<void>
+  }
+}
