+++ components.d.ts
@@ -1,0 +1,13 @@
+// generated by unplugin-vue-components
+// We suggest you to commit this file into source control
+// Read more: https://github.com/vuejs/vue-next/pull/3399
+
+declare module 'vue' {
+  export interface GlobalComponents {
+    FooterComponent: typeof import('./src/components/FooterComponent.vue')['default']
+    LoaderComponent: typeof import('./src/components/LoaderComponent.vue')['default']
+    ThemeButton: typeof import('./src/components/ThemeButton.vue')['default']
+  }
+}
+
+export { }
