+++ package.json
@@ -1,43 +1,43 @@
 {
   "name": "suveer-raina",
-  "version": "0.1.0",
   "private": true,
+  "version": "0.2.0",
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint"
+    "dev": "vite --port 3333 --open",
+    "build": "vue-tsc --noEmit && vite build",
+    "preview": "vite preview",
+    "typecheck": "vue-tsc --noEmit",
+    "lint": "eslint \"**/*.{vue,ts,js}\" --ignore-pattern \"/dist/\""
   },
   "dependencies": {
-    "core-js": "^3.21.1",
-    "register-service-worker": "^1.7.2",
-    "vue": "^3.2.31",
-    "vue-class-component": "^8.0.0-0",
-    "vue-meta": "^3.0.0-alpha.2",
-    "vue-router": "^4.0.12",
-    "vuex": "^4.0.2"
+    "@vueuse/head": "^0.7.5",
+    "pinia": "^2.0.11",
+    "vue": "^3.2.25"
   },
   "devDependencies": {
+    "@types/node": "^17.0.18",
     "@typescript-eslint/eslint-plugin": "^5.12.0",
     "@typescript-eslint/parser": "^5.12.0",
-    "@vue/cli-plugin-babel": "~5.0.1",
-    "@vue/cli-plugin-eslint": "~5.0.1",
-    "@vue/cli-plugin-pwa": "~5.0.1",
-    "@vue/cli-plugin-router": "~5.0.1",
-    "@vue/cli-plugin-typescript": "~5.0.1",
-    "@vue/cli-plugin-vuex": "~5.0.1",
-    "@vue/cli-service": "~5.0.1",
-    "@vue/compiler-sfc": "^3.2.31",
+    "@vitejs/plugin-vue": "^2.2.0",
     "@vue/eslint-config-prettier": "^7.0.0",
     "@vue/eslint-config-typescript": "^10.0.0",
-    "autoprefixer": "^10",
+    "autoprefixer": "^10.4.2",
     "eslint": "^8.9.0",
+    "eslint-config-prettier": "^8.4.0",
+    "eslint-plugin-html": "^6.2.0",
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-vue": "^8.4.1",
-    "postcss": "^8",
+    "postcss": "^8.4.6",
     "prettier": "^2.5.1",
     "sass": "^1.49.8",
-    "sass-loader": "^12.6.0",
-    "tailwindcss": "npm:@tailwindcss/postcss7-compat",
-    "typescript": "~4.5.5"
+    "tailwindcss": "^3.0.23",
+    "typescript": "^4.5.5",
+    "unplugin-vue-components": "^0.17.18",
+    "vite": "^2.8.0",
+    "vite-plugin-pages": "^0.20.2",
+    "vite-plugin-pwa": "^0.11.13",
+    "vite-plugin-vue-layouts": "^0.6.0",
+    "vue-router": "^4.0.12",
+    "vue-tsc": "^0.29.8"
   }
 }
