+++ public/_header
@@ -1,0 +1,3 @@
+/assets/*
+  cache-control: max-age=31536000
+  cache-control: immutable
