--- src/shims-vuex.d.ts
@@ -1,11 +1,0 @@
-// vuex.d.ts
-import { Store } from "vuex"
-import { Data } from "./store/interfaces"
-
-declare module "@vue/runtime-core" {
-  // provide typings for `this.$store`
-  interface ComponentCustomProperties {
-    $store: Store<Data>
-    // $store: Store<ColorsClass & LanguageState>
-  }
-}
