+++ src/store/interfaces.ts
@@ -1,0 +1,29 @@
+// Generated by https://quicktype.io
+
+export interface Data {
+	knowledge?: KnowledgeClass
+	colors?: ColorsClass
+}
+
+export interface ColorsClass {
+	primary: string
+	secondary: string
+	tertiary: string
+	contrast: string
+	dark: string
+	light: string
+	black: string
+	white: string
+}
+
+export interface KnowledgeClass {
+	language: { [key: string]: ObjClass }
+	"tools/frameworks": { [key: string]: ObjClass }
+	skills: { [key: string]: ObjClass }
+}
+
+export interface ObjClass {
+	lang: string
+	level: string | number
+	icon: string
+}
