--- src/store/interfaces/index.ts
@@ -1,29 +1,0 @@
-// Generated by https://quicktype.io
-
-export interface Data {
-  knowledge?: KnowledgeClass
-  colors?: ColorsClass
-}
-
-export interface ColorsClass {
-  primary: string
-  secondary: string
-  tertiary: string
-  contrast: string
-  dark: string
-  light: string
-  black: string
-  white: string
-}
-
-export interface KnowledgeClass {
-  language: { [key: string]: ObjClass }
-  "tools/frameworks": { [key: string]: ObjClass }
-  skills: { [key: string]: ObjClass }
-}
-
-export interface ObjClass {
-  lang: string
-  level: string | number
-  icon: string
-}
